requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
aiosqlite==0.19.0
python-telegram-bot>=21.0
aiohttp>=3.9.0

